+++ src/main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 #include "types.h"
-#include "engine.h"
 
 int main(int argc, char** argv) {
     return EXIT_SUCCESS;
